% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{ddp_url}
\alias{ddp_url}
\title{Generate a download url based on the DDP release key}
\usage{
ddp_url(key)
}
\arguments{
\item{key}{A DDP release key (character value)}
}
\value{
A url string (character value)
}
\description{
A helper function to map DDP release keys to download urls
}
\examples{

ddp_url("G19")
ddp_url("H15")
}
