% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{ddp_zipfile}
\alias{ddp_zipfile}
\title{Generate a DDP zip filename based on the DDP release key}
\usage{
ddp_zipfile(key, dir = ".")
}
\arguments{
\item{key}{A DDP release key (character value)}

\item{dir}{an optional directory name to include in the path}
}
\value{
A zipfile pathname (character value)
}
\description{
Map DDP release keys to download zip file names
}
\details{
Since DDP download urls are dynamic and do not include the filename, it is
neccessary to specify the filename when downloading from R. This function
replicates the file naming convention used by the DDP site.
}
\examples{

ddp_zipfile("G19")  # FRB_G19.zip
ddp_zipfile("H15")  # FRB_H15.zip)
}
