% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddptools.R
\docType{package}
\name{ddptools}
\alias{ddptools}
\alias{ddptools-package}
\title{ddptools: A package for processing Federal Reserve Board statistical releases}
\description{
FRB statistical releases are made available for download in SDMX format via
the Data Download Program (\url{https://www.federalreserve.gov/datadownload/}).
The size and structure of these files can make them somewhat inconvenient to
use. This package provides functions to make these data more accessible.
}
