% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{download_zip}
\alias{download_zip}
\title{Download data from the DDP website}
\usage{
download_zip(key, dir = ".", url = NULL, destfile = NULL, ...)
}
\arguments{
\item{key}{A valid DDP release key. If not specified, a valid download
\code{url} must be provided}

\item{dir}{A directory for the output file. The current directory
is the default}

\item{url}{A url string. If not specified the function will
determine the default download url based on the release \code{key}}

\item{destfile}{An output file path. If not specified the function
will determine the default filename based on the release \code{key}
and the \code{dir} option}

\item{...}{Any additional arguments are passed on to the \code{download.file}
function, which handles the download}
}
\value{
...
}
\description{
Given a release key, this function will download the zip file for the
corresponding statistical release from the DDP website. Alternatively,
provide a url and/or destination file.
}
\details{
Data Download Program (DDP) website: \url{https://www.federalreserve.gov/datadownload/}
}
\examples{

# save FRB_H15.zip in the current directory
download_zip(key = "H15")

# save FRB_H15.zip in the ./data-raw/ directory
download_zip(key = "H15", dir = "data-raw")
download_zip(key = "H15", destfile = "data-raw/FRB_H15.zip")
myurl <- "https://www.federalreserve.gov/datadownload/Output.aspx?rel=H15&filetype=zip"
download_zip(url = myurl, destfile = "data-raw/FRB_H15.zip")

}
