% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{extract_xml}
\alias{extract_xml}
\title{Extract SDMX data from DDP zip file}
\usage{
extract_xml(zipfile, ...)
}
\arguments{
\item{zipfile}{A file path for a DDP zip file}

\item{...}{Additional options for the \code{unzip} function}
}
\value{
name of the file that was extracted
}
\description{
DDP zip files contain multiple xml files. This function will selectively
extract the SDMX data file that contains the time series data. By default
this file follows the naming convention KEY_data.xml, where KEY is a DDP
release key.
}
\examples{

# Extract H15_data.xml in current directory
extract_xml("FRB_H15.zip")

# Extract H15_data.xml in data-raw directory
extract_xml("FRB_H15.zip", exdir = "data-raw")
extract_xml("FRB_H15.zip", files = "H15_data.xml", exdir = "data-raw")

}
