% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{lookup_obs_status}
\alias{lookup_obs_status}
\title{Lookup descriptions for valid SDMX obs_status codes}
\usage{
lookup_obs_status(code, standard = FALSE)
}
\arguments{
\item{code}{A character vector of obs_status codes for observations in an
SDMX time series}

\item{standard}{A logical value. If TRUE, nonstandard DDP codes will be filtered out
(mapped to \code{NA}).}
}
\value{
A character vector of descriptions for recognized obs_status codes
}
\description{
Map SDMX obs_status to descriptive text labels. Unrecognized codes
will be mapped to NA.
}
\details{
Valid codes are documented in the SDMX statistical standards:
\url{https://sdmx.org/?page_id=3215}. DDP data include additional nonstandard
codes. Set \code{standard = TRUE} to exclude these values.
}
