% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{lookup_release_name}
\alias{lookup_release_name}
\title{Map DDP release keys to FRB release names}
\usage{
lookup_release_name(key)
}
\arguments{
\item{key}{A character vector of DDP release keys}
}
\value{
A character vector
}
\description{
Map DDP release keys to release titles. Returns a complete list of known
release keys if no input argument is given
}
\examples{

lookup_release("G19")  # Consumer Credit (G.19)
lookup_release("H15")  # Selected Interest Rates (H.15)

lookup_release(c("PRATES", "H41" ))

lookup_release()       # returns a named vector with all releases

}
