% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{lookup_tif}
\alias{lookup_tif}
\title{Convert known SDMX frequency identifiers to tif codes for tis}
\usage{
lookup_tif(freq)
}
\arguments{
\item{freq}{A character vector of frequency identifier codes for an SDMX time
series.}
}
\value{
A character vector of tif codes
}
\description{
SDMX uses freq codes to describe the frequency of a time series.
This function translates freq codes into the tif codes used by
the tis package for the same purpose. This information is neccessary
to create a valid tis object
}
\examples{

lookup_tif(8)            # daily
lookup_tif(c("8", "9")   # daily, business
lookup_tif(c(9, -999))   # business, NA
lookup_tif()             # return all freq codes in a named vector
}
