% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ddp.R
\name{read_ddp}
\alias{read_ddp}
\title{Read an SDMX data file downloaded from the FRB DDP page}
\usage{
\code{read_ddp(file, pattern = NULL, skip = 0, n = Inf)}
}
\arguments{
\item{file}{A filename or a connection to an SDMX formatted file.}

\item{pattern}{A charcter vector of regular expressions used to filter input
by the \emph{series_name} attribute. Only matching series will be imported.}

\item{skip}{Number of series to skip starting from the beginning of the file.
If any \code{pattern} filters are provided, only matching series are
counted.}

\item{n}{Maximum number of series to import from the file. If any \code{pattern}
filters are provided, only matching series are counted.}
}
\value{
A list of \code{tis} objects
}
\description{
Read time series data and metadata stored in an SDMX formatted XML file, and
return a list of \code{tis} objects. Series metadata are stored as attributes
within each \code{tis} object.
}
\details{
Input files are read sequentially, line-by-line, to avoid storing the
entire XML file in memory. With the exception of the tis package, this
function is implemented using only built-in R functions.
}
\section{Note}{

This function is intended for use with the SDMX 1.0 formatted data files
provided by the Federal Reserve Board (refer to
\url{https://www.federalreserve.gov/datadownload/}). SDMX files from other
sources may not be processed correctly.
}

\examples{

library(tis)
library(ddptools)

# read only the first 5 series
db <- read_ddp("G19_data.xml", n = 5)

# skip the first 5 series, then read the next 10
db <- read_ddp("G19_data.xml", skip = 5, n = 10)

# include only matching series_name
db <- read_ddp("H15_data.xml", pattern = c("^RIF", "^RIM.*"))

# read all data
db <- read_ddp("PRATES_data.xml")

names(db)
attributes(db$RESBMS_N.D)
tail(db$RESBMS_N.D)

lm(RESBMS_N.D ~ lag(RESBME_N.D), db)

arima(db$RESBMS_N.D, c(0, 0, 1))

plot(db$RESBMS_N.D)

plot(
    x    = as.Date(ti(db$RESBMS_N.D)),
    y    = db$RESBMS_N.D,
    type = "s",
    main = attr(db$RESBMS_N.D, "long_description"),
    xlab = "Date",
    ylab = attr(db$RESBMS_N.D, "unit")
)

}
\seealso{
\link[tis]{tis}
}
